WARNING! This commit contains a migration reset. Renamed applicants to customers. Reset migrations.
