+++ .gitignore
@@ -67,3 +67,4 @@ target/
 /.pydevproject
 /.project
 
+**/local/
