+++ wildlifelicensing/apps/customers/apps.py
@@ -3,5 +3,5 @@ from __future__ import unicode_literals
 from django.apps import AppConfig
 
 
-class ApplicantsConfig(AppConfig):
-    name = 'applicants'
+class CustomersConfig(AppConfig):
+    name = 'customers'
