+++ wildlifelicensing/apps/customers/templates/customers_dashboard.html
@@ -6,5 +6,5 @@
 {% endblock %}
 
 {% block content %}
-    <h1>Applicants Dashboard</h1>
+    <h1>Customers Dashboard</h1>
 {% endblock %}
