+++ wildlifelicensing/apps/customers/views.py
@@ -2,4 +2,4 @@ from django.views.generic import TemplateView
 
 
 class DashboardView(TemplateView):
-    template_name = 'applicants_dashboard.html'
+    template_name = 'customers_dashboard.html'
