+++ wildlifelicensing/urls.py
@@ -8,7 +8,7 @@ urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url(r'^$', DashBoardView.as_view(), {'form': LoginForm}, name='home'),
     url(r'^accounts/', include('wildlifelicensing.apps.accounts.urls', namespace='accounts')),
-    url(r'^applicants/', include('wildlifelicensing.apps.applicants.urls', namespace='applicants')),
+    url(r'^customers/', include('wildlifelicensing.apps.customers.urls', namespace='customers')),
     url(r'^officers/', include('wildlifelicensing.apps.officers.urls', namespace='officers')),
     url(r'', include('social.apps.django_app.urls', namespace='social')),
 ]
